no message
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@264 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/include/opencv/cxmisc.h
@@ -202,7 +202,7 @@
 
 #define  CV_ZERO_OBJ(x)  memset((x), 0, sizeof(*(x)))
 
-#define  CV_DIM(static_array) (sizeof(static_array)/sizeof((static_array)[0]))
+#define  CV_DIM(static_array) ((int)(sizeof(static_array)/sizeof((static_array)[0])))
 
 #define  CV_UN_ENTRY_C1(worktype)           \
     worktype s0 = scalar[0]
